requirement.txt
numpy
opencv-python
sklearn
pandas
tensorflow
matploltlib
scipy
pywin32
speedtest-cli
pyqt5
pyqtgraph